pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "astra-assistants-api"
version = "0.1.0"
description = ""
authors = ["phact <estevezsebastian@gmail.com>"]

[[tool.poetry.packages]]
include = "openapi_server"

[tool.poetry.dependencies]
python = "^3.10.12"
aiofiles = "23.1.0"
aniso8601 = "7.0.0"
async-exit-stack = "1.0.1"
async-generator = "1.10"
chardet = "4.0.0"
click = "7.1.2"
dnspython = "2.1.0"
email-validator = "1.1.2"
fastapi = "0.95.2"
graphene = "2.1.8"
graphql-core = "2.3.2"
graphql-relay = "2.0.1"
httptools = "0.1.2"
idna = "2.10"
itsdangerous = "1.1.0"
markupsafe = "2.0.1"
orjson = "3.5.2"
promise = "2.3"
python-dotenv = "0.17.1"
python-multipart = "0.0.5"
pyyaml = "^6.0.1"
rx = "1.6.1"
six = "1.16.0"
starlette = "0.27.0"
typing-extensions = "4.8.0"
ujson = "4.0.2"
urllib3 = "1.26.5"
uvicorn = "0.13.4"
uvloop = "^0.19.0"
watchgod = "0.7"
websockets = "^12.0"
openai = "^1.1.2"
numpy = "^1.26.1"
cassandra-driver = "^3.28.0"
loguru = "^0.7.2"
arrow = "^1.3.0"
tenacity = "^8.2.3"
tiktoken = "^0.5.1"
docx2txt = "^0.8"
pypdf2 = "^3.0.1"
python-pptx = "^0.6.23"
gunicorn = "^21.2.0"
litellm = "1.20.6"
boto3 = "^1.29.6"
prometheus-fastapi-instrumentator = "^6.1.0"
google-cloud-aiplatform = "^1.38.0"
google-generativeai = "^0.3.1"
pytest = "^8.0.2"
streaming-assistants = "^0.14.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"

[tool.ruff.lint]
# Allow bare except (`E722`), ignore F-strings without placeholder (`F541`)
ignore = ["E722", "F541"]
# Add isort for linting
extend-select = ["I"]

# Ignore `E402` (import violations) in all `__init__.py` files
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]

[tool.black]
line-length = 88
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
skip = [
    '.eggs', '.git', '.hg', '.mypy_cache', '.nox', '.pants.d', '.tox',
    '.venv', '_build', 'buck-out', 'build', 'dist', 'node_modules', 'venv',
]
skip_gitignore = true
